Create the `useScreenHandler` hook
